+++ src/Category_Colors/CSS/V2_Views.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * The Events Calendar Category Colors
+ * The Events Calendar: Category Colors
  *
  * @author   Andy Fragen
  * @license  MIT
@@ -43,7 +43,7 @@ class V2_Views {
 		$css[] = ".tribe-common article.tribe_events_cat-{$slug} h3 a,";
 		$css[] = ".tribe-common article.tribe_events_cat-{$slug} h3 a:link,";
 		$css[] = ".tribe-common article.tribe_events_cat-{$slug} h3 a:visited,";
-		$css[] = "article.tribe-events-calendar-month__multiday-event.tribe_events_cat-{$slug} h3,";
+		$css[] = "article.tribe-events-calendar-month__multiday-event.tribe_events_cat-{$slug} h3";
 		self::echo_css( $css );
 	}
 
@@ -63,7 +63,7 @@ class V2_Views {
 		$css[] = "article.tribe-events-calendar-month-mobile-events__mobile-event.tribe_events_cat-{$slug} h3,";
 		$css[] = "article.tribe-events-calendar-day__event.tribe_events_cat-{$slug} h3,";
 		$css[] = "article.tribe-events-calendar-list__event.tribe_events_cat-{$slug} h3,";
-		$css[] = "article.tribe-events-calendar-latest-past__event.tribe_events_cat-{$slug} h3,";
+		$css[] = "article.tribe-events-calendar-latest-past__event.tribe_events_cat-{$slug} h3";
 		self::echo_css( $css );
 	}
 
@@ -77,6 +77,17 @@ class V2_Views {
 		$css[] = '.tribe-events .tribe-events-calendar-month__multiday-event-bar-inner,';
 		$css[] = '.tribe-events-calendar-month__multiday-event-wrapper';
 		$css[] = '{ background-color: #F7F6F6; }';
+		self::echo_css( $css );
+	}
+
+	/**
+	 * Make event spacer transparent.
+	 *
+	 * @return void
+	 */
+	public static function fix_spacer_background() {
+		$css[] = '.tribe-events-calendar-month__multiday-event-wrapper.tribe-events-calendar-month__multiday-event--empty';
+		$css[] = '{ background-color: transparent !important; }';
 		self::echo_css( $css );
 	}
 }